Create BeaconNodeButtons component
